Revert to milli epoch (#53)
This commit puts Aardvark back to milli epoch. Previous changes
moved to seconds epoch, but the logic that checks whether we're
doing an update looks to see that the new time is greater than
the old time. So everybody that upgrades won't get updates unless
we match the old format.